+++ packages/plugin-config/src/index.ts
@@ -1,21 +1,21 @@
 import * as path from 'path';
-import * as fse from 'fs-extra';
 import { IPlugin } from '@alib/build-scripts';
 
 const plugin: IPlugin = async (api): Promise<void> => {
   const { context, getValue, applyMethod } = api;
   const { command, rootDir } = context;
-
-  const configFile = `src/config.${getValue('PROJECT_TYPE')}`;
+  const configFilePattern = 'src/config';
 
   async function generateConfig() {
-    const filePath = path.join(rootDir,configFile);
+    const configFile = applyMethod('getSourceFile', configFilePattern, rootDir);
     const distPath =  path.join(getValue('TEMP_PATH'), 'config.ts');
     const templatePath = path.join(__dirname, './template/config.ts.ejs');
-    applyMethod('addRenderFile', templatePath, distPath, { hasConfig: fse.existsSync(filePath)});
+    applyMethod('addRenderFile', templatePath, distPath, { hasConfig: !!configFile });
+    const exportName = 'config, APP_MODE';
+    applyMethod('removeExport', exportName);
     applyMethod('addExport', {
       source: './config',
-      exportName: 'config, APP_MODE',
+      exportName,
       specifier: '{ config, APP_MODE }',
       importSource: '$$framework/config',
       exportMembers: ['config', 'APP_MODE'],
@@ -25,7 +25,7 @@ const plugin: IPlugin = async (api): Promise<void> => {
   generateConfig();
   if (command === 'start') {
     // watch folder config file is remove or added
-    applyMethod('watchFileChange', configFile, async (event: string) => {
+    applyMethod('watchFileChange', configFilePattern, async (event: string) => {
       if (event === 'unlink' || event === 'add') {
         await generateConfig();
       }