+++ packages/plugin-app-core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 1.4.10
+
+- [fix] globby pattern do not support path of win32
+
+## 1.4.9
+
+- [fix] ignore project type when get source file
+- [feat] register method of getSourceFile
+
 ## 1.4.8
 
 - [feat] add onPageNotFound and onUnhandledRejection types in ejs