+++ packages/plugin-config/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.8.2
+
+- [fix] get config file by plugin method
+
 ## 1.8.1
 
 - [fix] `userConfig` ts type in `config.ts.ejs`