+++ packages/plugin-webpack5/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.2.7
+
+- [fix] OOM when use filesystem cache with cnpm
+
+## 1.2.6
+
+- [fix] compatibility cache feature with webpack@5 of webpack-terser-plugin
+
 ## 1.2.5
 
 - [fix] compatible with win32 when delete original entry